requirements.txt
huggingface-hub<0.25.0
transformers==4.30.2
sentence-transformers==2.2.2
numpy<2.0.0
streamlit
pandas
pyyaml
scikit-learn
